Reduce multiple implementations to one
