+++ src/main/java/com/maddyhome/idea/vim/extension/multiplecursors/VimMultipleCursorsExtension.kt
@@ -34,7 +34,6 @@ import com.maddyhome.idea.vim.helper.SearchOptions
 import com.maddyhome.idea.vim.helper.endOffsetInclusive
 import com.maddyhome.idea.vim.helper.enumSetOf
 import com.maddyhome.idea.vim.helper.exitVisualMode
-import com.maddyhome.idea.vim.helper.findWordUnderCursor
 import com.maddyhome.idea.vim.helper.inVisualMode
 import com.maddyhome.idea.vim.helper.updateCaretsVisualAttributes
 import com.maddyhome.idea.vim.helper.userData
@@ -235,7 +234,7 @@ internal class VimMultipleCursorsExtension : VimExtension {
       val text = if (editor.inVisualMode) {
         primaryCaret.selectedText ?: return
       } else {
-        val range = findWordUnderCursor(editor, primaryCaret) ?: return
+        val range = injector.searchHelper.findWordNearestCursor(editor.vim, primaryCaret.vim) ?: return
         if (range.startOffset > primaryCaret.offset) return
         IjVimEditor(editor).getText(range)
       }
@@ -300,7 +299,8 @@ internal class VimMultipleCursorsExtension : VimExtension {
   }
 
   private fun selectWordUnderCaret(editor: Editor, caret: Caret): TextRange? {
-    val range = findWordUnderCursor(editor, caret) ?: return null
+    // TODO: I think vim-multiple-cursors uses a text object rather than the star operator
+    val range = injector.searchHelper.findWordNearestCursor(editor.vim, caret.vim) ?: return null
     if (range.startOffset > caret.offset) return null
 
     enterVisualMode(editor.vim)
