+++ src/main/java/com/maddyhome/idea/vim/helper/SearchHelperKt.kt
@@ -10,15 +10,10 @@ package com.maddyhome.idea.vim.helper
 
 import com.intellij.codeInsight.daemon.impl.DaemonCodeAnalyzerEx
 import com.intellij.codeInsight.daemon.impl.HighlightInfo
-import com.intellij.openapi.editor.Caret
 import com.intellij.openapi.editor.Editor
 import com.intellij.spellchecker.SpellCheckerSeveritiesProvider
-import com.maddyhome.idea.vim.api.getLineEndOffset
 import com.maddyhome.idea.vim.api.globalOptions
 import com.maddyhome.idea.vim.api.injector
-import com.maddyhome.idea.vim.common.TextRange
-import com.maddyhome.idea.vim.helper.CharacterHelper.charType
-import com.maddyhome.idea.vim.newapi.IjVimEditor
 import it.unimi.dsi.fastutil.ints.IntComparator
 import it.unimi.dsi.fastutil.ints.IntIterator
 import it.unimi.dsi.fastutil.ints.IntRBTreeSet
@@ -48,64 +43,6 @@ private fun containsUpperCase(pattern: String): Boolean {
     }
   }
   return false
-}
-
-/**
- * Find the word under the cursor or the next word to the right of the cursor on the current line.
- *
- * @param editor The editor to find the word in
- * @param caret  The caret to find word under
- * @return The text range of the found word or null if there is no word under/after the cursor on the line
- */
-fun findWordUnderCursor(editor: Editor, caret: Caret): TextRange? {
-  val vimEditor = IjVimEditor(editor)
-  val chars = editor.document.charsSequence
-  val stop = vimEditor.getLineEndOffset(caret.logicalPosition.line, true)
-
-  val pos = caret.offset
-  // Technically the first condition is covered by the second one, but let it be
-  if (chars.length == 0 || chars.length <= pos) return null
-
-  //if (pos == chars.length() - 1) return new TextRange(chars.length() - 1, chars.length());
-  var start = pos
-  val types = arrayOf(
-    CharacterHelper.CharacterType.KEYWORD,
-    CharacterHelper.CharacterType.PUNCTUATION
-  )
-  for (i in 0..1) {
-    start = pos
-    val type = charType(vimEditor, chars[start], false)
-    if (type == types[i]) {
-      // Search back for start of word
-      while (start > 0 && charType(vimEditor, chars[start - 1], false) == types[i]) {
-        start--
-      }
-    } else {
-      // Search forward for start of word
-      while (start < stop && charType(vimEditor, chars[start], false) != types[i]) {
-        start++
-      }
-    }
-
-    if (start != stop) {
-      break
-    }
-  }
-
-  if (start == stop) {
-    return null
-  }
-  // Special case 1 character words because 'findNextWordEnd' returns one to many chars
-  val end = if (start < stop &&
-    (start >= chars.length - 1 ||
-      charType(vimEditor, chars[start + 1], false) != CharacterHelper.CharacterType.KEYWORD)
-  ) {
-    start + 1
-  } else {
-    injector.searchHelper.findNextWordEnd(vimEditor, start, 1, false, false) + 1
-  }
-
-  return TextRange(start, end)
 }
 
 fun findMisspelledWords(
